Create tests for new Chatbot with ContextRetrieval
